refactor(based/models/mixers/transformer_vq): 重构 Transformer VQ 模型
- 更新导入路径，使用相对路径导入模块
- 移除 param_dtype 和 dtype 相关的代码
- 调整模型初始化和前向传播逻辑
- 优化 VQAttention 和 LearnableVQ 类的实现
- 更新实验配置文件，调整模型参数

+++ based/models/mixers/transformer_vq/nn/pe.py
@@ -2,7 +2,7 @@ import dataclasses
 import torch
 import torch.nn as nn
 
-from transformer_vq.nn.types import TransformerConfig
+from based.models.mixers.transformer_vq.nn.types import TransformerConfig
 
 
 def get_sinusoid_embs(length, width, lam, flip, start=0):
@@ -39,4 +39,4 @@ class ScaledSin(nn.Module):
         embs = get_sinusoid_embs(
             length=length, start=offset, width=self.d_model, lam=self.pe_lam, flip=False
         )  # UD
-        return (self.scale * embs).to(self.dtype)
+        return (self.scale * embs)
